# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Canopus
  description: A lightweight encryption service inspired by HashiCorp Vault.
  version: 0.1.0

servers:
  - url: http://localhost:2107
    description: Local development

security:
  - ApiKeyAuth: []

tags:
  - name: Crypto
    description: Encrypt and decrypt data
  - name: Keys
    description: Key management and rotation

paths:
  /encrypt:
    post:
      summary: Encrypt a value
      tags: [Crypto]
      parameters:
        - name: key_version
          in: query
          required: false
          description: Key version to use for encryption. Defaults to the latest key.
          schema:
            type: integer
            example: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncryptRequest'
            examples:
              string:
                summary: Encrypt a string
                value:
                  plaintext: "hello world"
              object:
                summary: Encrypt a JSON object
                value:
                  plaintext:
                    user: waldo
                    role: admin
      responses:
        '200':
          description: Data encrypted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncryptResponse'
              example:
                status: success
                data:
                  ciphertext: "canopus:v0:f:gAAAAABp..."
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'

  /decrypt:
    post:
      summary: Decrypt a ciphertext
      tags: [Crypto]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecryptRequest'
            example:
              ciphertext: "canopus:v0:f:gAAAAABp..."
      responses:
        '200':
          description: Data decrypted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecryptResponse'
              examples:
                string:
                  summary: Decrypted string
                  value:
                    status: success
                    data:
                      plaintext: "hello world"
                object:
                  summary: Decrypted JSON object
                  value:
                    status: success
                    data:
                      plaintext:
                        user: waldo
                        role: admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /rotate-key:
    post:
      summary: Rotate the active encryption key
      description: >
        Generates a new encryption key and appends it as the active version.
        All previous key versions are retained so existing ciphertexts remain
        decryptable. Requires the `X-Canopus-Rotate-Token` header when
        `CANOPUS_API_KEY` is configured. Rate limited to 5 requests per hour
        per IP. Can be disabled entirely via `CANOPUS_ROTATION_ENABLED=false`.
      tags: [Keys]
      security:
        - ApiKeyAuth: []
        - RotateKeyAuth: []
      parameters:
        - name: X-Canopus-Rotate-Token
          in: header
          required: false
          description: >
            Separate key for rotation operations. Required when `CANOPUS_API_KEY`
            is set. Falls back to `CANOPUS_API_KEY` if `CANOPUS_ROTATE_KEY` is
            not configured.
          schema:
            type: string
      responses:
        '200':
          description: Key rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateKeyResponse'
              example:
                status: success
                data:
                  message: Key rotated successfully
                  key_version: 6
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/RotationDisabled'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '501':
          $ref: '#/components/responses/RotationNotSupported'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  schemas:
    EncryptRequest:
      type: object
      required: [plaintext]
      properties:
        plaintext:
          description: >
            Value to encrypt. Supported types: string, number, object, array.
          oneOf:
            - type: string
            - type: number
            - type: object
              additionalProperties: true
            - type: array
              items: {}

    EncryptResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            ciphertext:
              type: string
              description: >
                Encrypted value in the format `canopus:v{version}:{cipher}:{token}`.
                Both the key version and cipher alias are embedded — clients never need to track them.
                Cipher aliases: f (Fernet/AES-128-CBC), g (AES-256-GCM), c (AES-256-CBC).
              example: "canopus:v0:f:gAAAAABp..."

    DecryptRequest:
      type: object
      required: [ciphertext]
      properties:
        ciphertext:
          type: string
          description: Ciphertext produced by the `/encrypt` endpoint.
          example: "canopus:v0:gAAAAABp..."

    DecryptResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            plaintext:
              description: Decrypted value. Type matches the original value passed to `/encrypt`.
              oneOf:
                - type: string
                - type: number
                - type: object
                  additionalProperties: true
                - type: array
                  items: {}

    RotateKeyResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          properties:
            message:
              type: string
              example: Key rotated successfully
            key_version:
              type: integer
              description: The version number of the newly active key.
              example: 6

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        data:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error description.
            code:
              type: string
              description: Machine-readable error code.
              enum:
                - MISSING_FIELD
                - INVALID_KEY_VERSION
                - UNSUPPORTED_TYPE
                - INVALID_CIPHERTEXT
                - ROTATION_NOT_SUPPORTED
                - ROTATION_DISABLED
                - RATE_LIMIT_EXCEEDED
                - UNAUTHORIZED
                - INTERNAL_ERROR

  responses:
    BadRequest:
      description: Invalid request — missing or malformed input.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: 'Missing "plaintext" key in request body'
              code: MISSING_FIELD

    UnprocessableEntity:
      description: Input is well-formed but semantically invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: Unsupported type for "plaintext" value
              code: UNSUPPORTED_TYPE

    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: Missing X-Canopus-Token header
              code: UNAUTHORIZED

    RotationDisabled:
      description: Key rotation is disabled via CANOPUS_ROTATION_ENABLED=false.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: Key rotation is disabled
              code: ROTATION_DISABLED

    RateLimitExceeded:
      description: Rate limit exceeded — maximum 5 rotation requests per hour per IP.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: Rate limit exceeded
              code: RATE_LIMIT_EXCEEDED

    RotationNotSupported:
      description: The active storage backend does not support key rotation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: The active storage backend does not support key rotation.
              code: ROTATION_NOT_SUPPORTED

    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data:
              message: An unexpected error occurred
              code: INTERNAL_ERROR

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Canopus-Token
      description: >
        API key for authentication. Set via the CANOPUS_API_KEY environment variable.
        When CANOPUS_API_KEY is not configured, authentication is disabled (development mode).

    RotateKeyAuth:
      type: apiKey
      in: header
      name: X-Canopus-Rotate-Token
      description: >
        Separate key for /rotate-key operations. Set via CANOPUS_ROTATE_KEY.
        Falls back to CANOPUS_API_KEY if CANOPUS_ROTATE_KEY is not configured.